MOVE openshift version and csv to capAudit
On this commit:

ocp version and csv are being moved to capAudit data structure to make
that data available through all the lifecycle of a single audit. This
allows for passing the whole capAudit object to reporting with much more
data in context then before therefore enabling more sophisticated
reports.

Signed-off-by: acmenezes <adcmenezes@gmail.com>