handle errors in kubeConfig() function
This change first and foremost aims to provide more descriptive errors
when an error in encountered while obtaining the kube config.

In addition, it moves error handling when obtaining a kube config to the kubeConfig() function
to alleviate the need to handle the errors where the kube config where
the function is called.

Closes #159

Signed-off-by: Edmund Ochieng <ochienged@gmail.com>