return unique subscription names in GetSubscriptionsData method
Update the GetSubscriptionsData() method to return all install plans supported by
the operator to be returned to the user as unique subscription data
objects.

In addition, update the subscription name to include the install mode
which would allow each subscription name to be unique.

This changes are to prepare for testing of all install modes supported
by an operator.

Signed-off-by: Edmund Ochieng <ochienged@gmail.com>