REMOVE unnecessary exported methods from capability
On this commit all capAudit objects and methods were changed to lower
case in order to be non exported. The capAudit interface was removed
also to prevent confusion once the only interface to this package is
CapAuditor and not capAudit. CapAuditor now should be upper case and
exported. More on the documentation commit.

Signed-off-by: acmenezes <adcmenezes@gmail.com>