REMOVE unnecessary code from Auditor.go
We don't need a builder for this simple object and buildWorkQueue can be
run as the first step of an audit. That makes the code simpler and
easier to understand. All flag data interesting to capability package
can be passed by using the CapAuditor object and add to it as needed. It
makes CapAuditor the single entry point of flag data to the capabiltiy
package.

Signed-off-by: acmenezes <adcmenezes@gmail.com>