Fix panic introduced by context param
When the context parameter was added to the audit methods, it caused
a panic. This was not caught because of the use of reflection. This
will be a short-term fix for this panic. The long-term fix is to get
rid of the reflection and instead opt for a standard interface that
each audit can implement.

Fixes #220

Signed-off-by: Brad P. Crochet <brad@redhat.com>