Refactor audits to remove reflection
The reflection being used was error-prone, and changes to the audit
signature resulted in a panic that should have been caught at compile
time. But since reflection was being used, it did not allow for that
type of error to be caught. This makes the audit calls more type-safe
and changes can be caught before runtime.

Signed-off-by: Brad P. Crochet <brad@redhat.com>