Allow users to provide additional Custom Resources
This commit adds the '--extra-cr-directory' flag to the check command.
The expected directory structures is:

custom_resources_directory/
├── package_name1
│   ├── manifest_file1.json
│   └── manifest_file2.yaml
└── package_name2
    ├── manifest_file1.json
    └── manifest_file2.yaml

Additional CRs can be specified for multiple packages by creating
multiple subdirectories. The name of the subdirectory should correspond
to the name of the package the manifest files are to be be deployed
with.

Multiple manifest files can be specified for each package. Both JSON and
YAML formats are accepted.

The additional CRs are merged with the ALMExamples, if any.

Fixes #130