Add an option to pass an afero filesystem to audits
To make audits easier to test, pass in a filesystem to be used for
writing files. This will allow the filesystem to be changed out with
for instance a memmap filesystem so it can be thrown away after a
test is run without having to write actual files to the physical
filesystem.

Fixes #298

Signed-off-by: Brad P. Crochet <brad@redhat.com>